pyproject.toml
[project]
name = "hasgeek-workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.1.1",
    "firecrawl-py>=4.3.6",
    "ipykernel>=6.30.1",
    "ipython>=9.6.0",
    "jupyter>=1.1.1",
    "jupyterlab>=4.4.9",
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.1.12",
    "langchain-mcp-adapters>=0.1.11",
    "langchain-openai>=0.3.34",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.8",
    "langgraph-cli[inmem]>=0.4.2",
    "langgraph-prebuilt>=0.6.4",
    "langgraph-sdk>=0.2.9",
    "matplotlib>=3.10.6",
    "mcp[cli]>=1.16.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.2",
    "yfinance>=0.2.66",
]
